Remove the option for enableWatcher
I guess I've realized that I want this for a reason, and the narrower of
a focus I have, the better it works.

If you want my _original_ use case, which is simply setting up window
arrangements, check out hs.window.layout and hs.chooser and roll your
own. ¯\(°_o)/¯