Make Toggl's current timer async
I figured out that the main reason that I would get slow response
launching the Headspace chooser is that it was blocking on getting the
current timer.

I probably need to refactor, but this means that it starts the HTTP
request when loaded, but continues to show whatever until the HTTP
request returns. Also helps on slow connections.